add docker pull section
